1.18.38.218 - port in propertiesFile and fixed popUp-Windows.
If the advanced settings are exported, then the port is included in it.
The same with the import.
The ipAddress and port in the DockPanel -bottom is updated when the port changes.

If the application was in fullscreen, then the popUp-Windows were always opened in fullscreen too.
Now: the popUp-Windows have a fixed size and cannot be moved or resized.