1.18.43.224 - Server stopped improved. only properties can be imported.
If the teacher stops the server, then all the student are red and every studentServer is shut down. This results to: client lost connection. Because the connection is lost, the student will be logged out.

Applied an FileExtensionFilter for the advanced-settings-import.
Only *.properties-files can be imported.

If the user presses ESC again then the "are you sure you want to close it"-Window will be closed.